[feat] implement toString

+++ rvdecoderdb/src/parser/BitValue.scala
@@ -10,9 +10,9 @@ object BitValue {
         new BitValue(
           bit.toInt,
           value match {
-            case s"0b$bstr" => BigInt(bstr, 2).toInt
-            case s"0x$xstr" => BigInt(xstr, 16).toInt
-            case dstr       => dstr.toInt
+            case s"0b$bstr" => BigInt(bstr, 2)
+            case s"0x$xstr" => BigInt(xstr, 16)
+            case dstr       => BigInt(dstr)
           }
         )
       )
@@ -20,6 +20,6 @@ object BitValue {
   }
 }
 
-class BitValue(val bit: Int, val value: Int) extends Token {
+class BitValue(val bit: BigInt, val value: BigInt) extends Token {
   override def toString: String = s"$bit=$value"
 }
